requirements.txt
Flask
flask_script
Flask-Bootstrap
Flask_SQLAlchemy
requests
pandevice
flask.mongoengine
